+++ vars/isStaticWorker.groovy
@@ -16,46 +16,14 @@
 // under the License.
 
 /**
- Return the name of the Operating system based on the labels of the Node.
+  Whether the existing worker is a static one
+  whenTrue(isStaticWorker(labels: 'linux&&immutable')) {
+    echo "I'm a static worker"
+  }
 
- def os = nodeOS()
-
+  TODO: as soon as ARM and MacOS are ephemerals then we need to change this method
 */
-def call() {
-  def labels = env.NODE_LABELS?.toLowerCase()
-  def matches = []
-
-  if (isLinux(labels) || (isArm() && !isDarwin(labels))) {
-    matches.add('linux')
-  }
-
-  if (isWindows(labels)) {
-    matches.add('windows')
-  }
-
-  if (isDarwin(labels)) {
-    matches.add('darwin')
-  }
-
-  if(matches.size() == 0){
-    error("Unhandled OS name in NODE_LABELS: " + labels)
-  }
-
-  if(matches.size() > 1){
-    error("Labels conflict OS name in NODE_LABELS: " + labels)
-  }
-
-  return matches[0]
-}
-
-def isLinux(labels){
-  return labels.contains('linux')
-}
-
-def isDarwin(labels){
-  return labels.contains('darwin') || labels.contains('macos')
-}
-
-def isWindows(labels){
-  return labels.contains('windows')
+def call(Map args=[:]){
+  def labels = args.containsKey('labels') ? args.labels : error("isStaticWorker: labels parameter is required.")
+  return (labels?.contains('arm') || labels?.contains('macosx') || labels?.contains('metal'))
 }